test(VideoPlayer): Add testing for format embed along with refactor reusable
