+++ package.json
@@ -54,6 +54,7 @@
     "@storybook/addon-knobs": "^6.0.28",
     "@storybook/addons": "^6.0.28",
     "@storybook/react": "^6.0.28",
+    "@testing-library/react": "^11.2.3",
     "@types/babel__core": "^7.1.3",
     "@types/enzyme": "^3.10.3",
     "@types/jest": "^24.0.23",
